fix transport shutdown hanging issue
This fixes the issue where the server isn't shutting down properly and
is relying on a 10s timeout in fx to shutdown.
